Builds and runs on device
main.rs was adapted from a recommended example in the stm32f4xx_hal crate
which I copied in and built. I adjusted its pin and clock to match the 1Bitsy
and it loaded and ran.

The rcc configuration is limited and takes fewer steps.

Prior behavior:

During clock setup, specifically the freeze() step, a register write was being
performed to 0x40023808. Working backwards, this address seems to be SPI2_SR
which is the status register for the second SPI host. Presumably this register
is read-only and changes only on read or in response to other actions.
Additionally, it was writing a 16-bit value that was setting 3 bits in the
upper 8 bits, which in the spec for this chip are reserved.

No idea why that was happening. This dodges the issue.